+++ MangAdventure/utils/images.py
@@ -1,7 +1,9 @@
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.conf import settings
 from os import path, remove, makedirs
+from constance import config
 from zipfile import ZipFile
+from shutil import rmtree
 from sys import getsizeof
 from io import BytesIO
 from PIL import Image
@@ -23,6 +25,16 @@ def thumbnail(obj, max_size=100):
 
 def unzip(obj):
     counter = 0
+    dir_path = path.join(
+        'series',
+        obj.series.slug,
+        str(obj.volume),
+        '%g' % obj.number
+    )
+    full_path = path.join(settings.MEDIA_ROOT, dir_path)
+    if path.exists(full_path):
+        rmtree(full_path)
+    makedirs(full_path)
     zip_file = ZipFile(obj.file)
     name_list = zip_file.namelist()
     for name in sort.natural_sort(name_list):
@@ -31,20 +43,10 @@ def unzip(obj):
         counter += 1
         data = zip_file.read(name)
         filename = '%03d%s' % (counter, path.splitext(name)[-1])
-        file_path = path.join(
-            'series',
-            obj.series.slug,
-            str(obj.volume),
-            str(obj.number),
-            filename
-        )
-        full_path = path.join(settings.MEDIA_ROOT, file_path)
-        if not path.exists(path.dirname(full_path)):
-            makedirs(path.dirname(full_path))
-        if path.exists(full_path):
-            remove(full_path)
+        file_path = path.join(dir_path, filename)
         image = Image.open(BytesIO(data))
-        image.save(full_path, optimize=True, quality=100)
+        image.save(path.join(full_path, filename),
+                   quality=100, optimize=config.COMPRESS_PAGES)
         obj.pages.create(number=counter, image=file_path)
     zip_file.close()
     # TODO: option to keep zip file